# swagger.yaml
openapi: 3.0.0
info:
  title: Zodiac Sign API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server 
  - url: https://heroku-deploy-zodiac.herokuapp.com/api/docs/
    description: Heroku remote server
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token

  schemas:
    Sign:
      type: object
      properties:
        zodiac:
          type: string
        number:
          type: number
        traits:
          type: string
        element:
          type: string
        month:
          type: string
        inLuck:
          type: boolean
    
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string

paths:

  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      
      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string


  /sign:
    get:
      summary: Retrieves a list of zodiac signs
      description: Retrieves a list of products as JSON objects.
      responses:
        200:
          description: A list of zodiac sign JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sign"
              
                  
    
    post:
      summary: Create a new zodiac sign
      description: Create a new zodiac sign
      security:
        - ApiKeyAuth: []      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sign"

      responses:
        201:
          description: Zodiac sign created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sign"
  
  /sign/{id}:
    get:
      summary: Specific zodiac sign
      description: Retrieves a specific zodiac sign based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A sign in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sign"
   
    put:
      summary: Updates a specific zodiac sign
      description: Updates a specific zodiac sign based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Sign"

      responses:
        201:
          description: Zodiac sign updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signs"

    delete:
      summary: Deletes a specific zodiac sign
      description: Deletes a specific zodiac sign based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string      

      responses:
        201:
          description: Zodiac sign deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signs"